AgriAdvisior/pyproject.toml
[project]
name = "agriadvisior"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.27.5",
    "googlemaps>=4.10.0",
    "requests>=2.33.0",
]
